chore: Moved base.endpoint.mts and endpoint.registry.mts to core/endpoint/.
