+++ packages/architectura/src/core/server/server.mts
@@ -8,11 +8,12 @@ import { PortsEnum } from "../../definition/enum/ports.enum.mjs";
 
 import { getConstructorOf } from "../../definition/type/get-constructor-of.type.mjs";
 
-import { EndpointRegistry } from "../../service/dispatcher/endpoint.registry.mjs";
 
 import { FileSystemService } from "../../service/file-system/file-system.service.mjs";
 
 import { LoggerProxy } from "../../service/logger/logger.proxy.mjs";
+
+import { EndpointRegistry } from "../endpoint/endpoint.registry.mjs";
 
 import { ExecutionContext } from "../execution-context/execution-context.mjs";
 
@@ -26,11 +27,12 @@ import type { ServerConfigurationType } from "../../definition/type/server-confi
 
 import type { ServerInstantiationType } from "../../definition/type/server-instantiation.type.mjs";
 
-import type { BaseEndpoint } from "../../endpoint/base.endpoint.mjs";
 
 import type { BasePostHook } from "../../hook/base.post-hook.mjs";
 
 import type { BasePreHook } from "../../hook/base.pre-hook.mjs";
+
+import type { BaseEndpoint } from "../endpoint/base.endpoint.mjs";
 
 class Server
 {
